requirements.txt
Flask
Flask-RESTful
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-Reuploaded
flask-marshmallow
flask-migrate
SQLAlchemy
marshmallow
marshmallow-sqlalchemy
uwsgi
psycopg2
python-dotenv